requirements.txt
asgiref==3.2.10
Django==3.1.2
pytz==2020.1
sqlparse==0.4.1

pip~=20.2.3
numpy~=1.19.2
setuptools~=41.2.0
djangorestframework~=3.12.1